+++ Modules/Alerts.lua
@@ -28,41 +28,6 @@ function addon:UpdateAlert(alertType, alert, input)
     end
 end
 
-function addon:PreviewBarAlert(input)
-    -- local barIDName = format("%s %s", L["Bar"], 1)
-    -- local progressCount = self:GetDBValue("global", "settings.alerts.bar.preview.count")
-    -- local progressTotal = self:GetDBValue("global", "settings.alerts.bar.preview.total")
-    -- local alertType = self:GetDBValue("global", "settings.alerts.bar.preview.alertType")
-
-    -- local alertInfo = {
-    --     progressCount = progressCount,
-    --     progressTotal = progressTotal,
-    --     barIDName = barIDName,
-    --     barNameLong = format("%s (%s)", barIDName, L["My Bar Name"]),
-    --     progressColor = (progressCount == progressTotal and alertType ~= "lost") and "|cff00ff00" or alertType == "complete" and "|cffffcc00" or alertType == "lost" and "|cffff0000",
-    -- }
-
-    -- if alertType == "complete" and progressCount == 0 then
-    --     return ""
-    -- elseif progressTotal < progressCount then
-    --     return format("%s: %s", L.Error, L.InvalidBarPreviewTotal)
-    -- else
-    --     local func, err = loadstring("return " .. (input or self:GetDBValue("global", "settings.alerts.bar.format.progress") or ""))
-    --     local success, error = pcall(func, addon, alertInfo)
-
-    --     -- Syntax error
-    --     if not success then
-    --         return L.InvalidSyntax(error), true
-    --     elseif not assert(func)() then 
-    --         return L.InvalidFunction, true
-    --     elseif not assert(func)()(alertInfo) or type(assert(func)()(alertInfo)) ~= "string" then
-    --         return L.InvalidReturn, true
-    --     else
-    --         return "|cffffffff" .. assert(func)()(alertInfo) .. "|r"
-    --     end
-    -- end
-end
-
 function addon:PreviewAlert(alertType, input, info)
     local alertInfo
 
@@ -79,6 +44,10 @@ function addon:PreviewAlert(alertType, input, info)
             barIDName = barIDName,
             barNameLong = format("%s (%s)", barIDName, L["My Bar Name"]),
             progressColor = (progressCount == progressTotal and alertType ~= "lost") and "|cff00ff00" or alertType == "complete" and "|cffffcc00" or alertType == "lost" and "|cffff0000",
+            difference = {
+                sign = alertType == "lost" and "-" or "+",
+                color = alertType == "lost" and "|cffff0000" or "|cff00ff00",
+            },
         }
     elseif alertType == "button" then
         local objective = self:GetDBValue("global", "settings.alerts.button.preview.objective")
@@ -198,6 +167,6 @@ function addon:SendAlert(bar, alertType, alert, alertInfo, soundID, isTracker)
         local newCompletion = oldCount < objective and newCount > objective
         local lostCompletion = oldCount > objective and newCount < objective
 
-        -- bar:AlertProgress("progress", (newCompletion and "complete") or (lostCompletion and "lost"))
+         bar:AlertProgress("progress", (newCompletion and "complete") or (lostCompletion and "lost"))
     end
 end